% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_diff_mat_function.R
\name{abs_diff_mat}
\alias{abs_diff_mat}
\title{abs_diff_mat}
\usage{
abs_diff_mat(DF, attrname)
}
\arguments{
\item{DF}{Dataframe of node attribute}

\item{attrname}{names of the attribute from the dataframe to create the matrix for.}
}
\value{
Absolute difference matrix
}
\description{
This takes a dataframe of node attributes and convert one into a absolute difference matrix
}
