% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELEnet16_data.R
\docType{data}
\name{ELEnet16}
\alias{ELEnet16}
\title{Electrical Automotive Goods 2016 Network}
\usage{
ELEnet16
}
\description{
Electrical Automotive Goods 2016 Network. Electrical automotive goods category as defined by Amighini & Gogoni (2014)
}
\references{
Amighini, A. and Gorgoni, S. (2014) The International Reorganisation of Auto Production, The World Economy, 37(7), pp. 923–952.
}
\keyword{datasets}
