% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELEnetList_data.R
\docType{data}
\name{ELEnetList}
\alias{ELEnetList}
\title{List of Electrical Automotive Goods Networks (2006-2016)}
\usage{
ELEnetList
}
\description{
List of Electrical Automotive Goods Networks for 2006 - 2016. Electrical automotive goods category as defined by Amighini & Gogoni (2014)
}
\references{
Amighini, A. and Gorgoni, S. (2014) The International Reorganisation of Auto Production, The World Economy, 37(7), pp. 923–952. (list of igraph objects)
}
\keyword{datasets}
