% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GF dataframe_function.R
\name{GF_dataframe}
\alias{GF_dataframe}
\title{Gould & Fernandez (1989) Brokerage Roles Dataframe}
\usage{
GF_dataframe(gs, attrname)
}
\arguments{
\item{gs}{International Trade Network - igraph object}

\item{attrname}{Attribute name}
}
\value{
Table of Gould-Fernandez results (dataframe)
}
\description{
This function calculates the Gould & Fernandez (1989) brokerage roles for an attribute
}
\examples{
require(igraph)
##Create random network (igraph object)
gs<-erdos.renyi.game(75,0.05,directed = TRUE)

##Add vertex names
V(gs)$name<-1:vcount(gs)

## Add an attribute
V(gs)$letters<- rep(LETTERS[1:5],15)

##Calculate the GF results
GF_results<-GF_dataframe(gs,"letters")

}
