% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GF_path_list_function.R
\name{GF_path_list}
\alias{GF_path_list}
\title{Gould & Fernandez (1989) Brokerage Roles path list}
\usage{
GF_path_list(gs, attrname)
}
\arguments{
\item{gs}{International Trade Network - igraph object}

\item{attrname}{Attribute name}
}
\value{
List of path dataframes: Coordinator, Representative, Gatekeeper, Consultant & Liaison
}
\description{
This function calculates returns all the Gould & Fernandez (1989) brokerage paths for an attribute
}
\examples{
require(igraph)
##Create random network (igraph object)
gs<-erdos.renyi.game(75,0.05,directed = TRUE)

##Add vertex names
V(gs)$name<-1:vcount(gs)

## Add an attribute
V(gs)$letters<- rep(LETTERS[1:5],15)

##Calculate the GF results
GF_PATH<-GF_path_list(gs,"letters")

}
