% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isEmpty_function.R
\name{isEmpty}
\alias{isEmpty}
\title{isEmpty}
\usage{
isEmpty(x)
}
\arguments{
\item{x}{Data}
}
\value{
NA or the data
}
\description{
This function check whether data is numeric(0) and give returns an NA if this is true and the value of the data otherwise.
}
