% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ITN_map_plot_function.R
\name{ITN_map_plot}
\alias{ITN_map_plot}
\title{ITN plot on world map}
\usage{
ITN_map_plot(gs)
}
\arguments{
\item{gs}{International Trade Network - igraph object}
}
\value{
Plot of the ITN on world map
}
\description{
This function plots the international trade network on a world map
}
\examples{
\donttest{
require(maps)
##Load the ITN
data(ELEnet16)

## Plot ITN on map - node size based on outdegree
ITN_map_plot(ELEnet16)
}
}
