% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ITNblock_plot_function.R
\name{ITNblock_plot}
\alias{ITNblock_plot}
\title{ITN Blockmodel & Structural Equivalence}
\usage{
ITNblock_plot(gs, LABEL)
}
\arguments{
\item{gs}{International Trade Network - igraph object}

\item{LABEL}{Should labels be present - TRUE/FALSE}
}
\value{
Network Plot - nodes coloured based on block membership
}
\description{
This function calculates block membership for ITN and strucutral equivalence between countries
}
\examples{
\dontrun{
##Create random International Trade Network (igraph object)
ITN<-erdos.renyi.game(75,0.05,directed = TRUE)

##Blockmodel plot
ITNblock_plot(ITN,FALSE)

}
}
