% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ITNcluster_function.R
\name{ITNcluster}
\alias{ITNcluster}
\title{ITN Cluster}
\usage{
ITNcluster(gs)
}
\arguments{
\item{gs}{International Trade Network - igraph object (with region attribute)}
}
\value{
Cluster object containing various cluster memebrship and correlation results
}
\description{
This function calculates cluster membership for ITN and compares with regional groupings
}
\examples{
\dontrun{
##Create random International Trade Network (igraph object)
ITN<-erdos.renyi.game(75,0.05,directed = TRUE)

##Cluster Analysis
CLU<-ITNcluster(ITN)

}
}
