% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ITNdegdist_function.R
\name{ITNdegdist}
\alias{ITNdegdist}
\title{ITN Degree Distribution}
\usage{
ITNdegdist(gs)
}
\arguments{
\item{gs}{International Trade Network - igraph object}
}
\value{
Panel of ITN degree distribtuion plots
}
\description{
This function plots the ITN degree distribtuion
}
\examples{
\dontrun{
##Create random International Trade Network (igraph object)
ITN<-erdos.renyi.game(75,0.05,directed = TRUE)

##Plot degree distribution
ITNdegdist(ITN)

}
}
