% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ITNplotset_function.R
\name{ITNplotset}
\alias{ITNplotset}
\title{ITN Plots}
\usage{
ITNplotset(gs)
}
\arguments{
\item{gs}{International Trade Network - igraph object}
}
\value{
Panel of ITN plots
}
\description{
This function plots the ITN for a quick inspection
}
\examples{
\dontrun{
##Create random International Trade Network (igraph object)
ITN<-erdos.renyi.game(75,0.05,directed = TRUE)

##Plot set of network visualisations
ITNplotset(ITN)

}
}
