% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ITNproperties_function.R
\name{ITNproperties}
\alias{ITNproperties}
\title{ITN Properties}
\usage{
ITNproperties(gs)
}
\arguments{
\item{gs}{International Trade Network - igraph object}
}
\value{
Table of centrality results (dataframe)
}
\description{
This function calculates network level properties of the International Trade Network (ITN)
}
\examples{
\dontrun{
##Create random International Trade Network (igraph object)
ITN<-erdos.renyi.game(75,0.05,directed = TRUE)

##Add edge weights
E(ITN)$weight<-runif(ecount(ITN), 0, 1)

##Calculate the network properties
ITNPROP<-ITNproperties(ITN)

}
}
