% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_trade_network_function.R
\name{make_trade_network}
\alias{make_trade_network}
\title{make_trade_network}
\usage{
make_trade_network(DF, threshold, cutoff)
}
\arguments{
\item{DF}{Dataframe of trade data downloaded (potentially using the comtradr package)}

\item{threshold}{Apply a threshold - TRUE, Extract the backbone - FALSE}

\item{cutoff}{Threshold - cutoff level, Backbone - significance level}
}
\value{
International Trade Network - igraph object
}
\description{
This function takes (import) trade data and cleans it and transforms it into a network.
This function can be applied to trade data downloaded from UN Comtrade (download csv and read into R as a dataframe), or any other trade data. You just make sure it has the following column names:
reporter_iso, partner_iso and edge_weight. Some dataformats may have different names. Also - it is important to note that this function is for import data.
}
