% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_degree_distribution_function.R
\name{plot_degree_distribution}
\alias{plot_degree_distribution}
\title{Plot Degree Distribution}
\usage{
plot_degree_distribution(graph, a)
}
\arguments{
\item{graph}{igraph object}

\item{a}{mode - "in","out","all}
}
\value{
Panel of ITN degree distribution plots
}
\description{
This function plots degree distribution for any graph
}
\examples{
require(igraph)
##Create random International Trade Network (igraph object)
ITN<-erdos.renyi.game(75,0.05,directed = TRUE)

##Plot out degree distribution
plot_degree_distribution(ITN,"in")
}
