% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data_frame_function.R
\name{process_data_frame}
\alias{process_data_frame}
\title{process_data_frame}
\usage{
process_data_frame(DF)
}
\arguments{
\item{DF}{edgelist with sender, receiver and value columns.}
}
\value{
Country neteork - igraph object
}
\description{
This function tan edge lsit of countries, and adds attribute data. It creates a network with attributes.
}
