% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/receiver_mat_function.R
\name{receiver_mat}
\alias{receiver_mat}
\title{receiver_mat}
\usage{
receiver_mat(DF, attrname)
}
\arguments{
\item{DF}{Dataframe of node attribute}

\item{attrname}{names of the attribute from the dataframe to create the matrix for.}
}
\value{
Receiver matrix
}
\description{
This takes a dataframe of node attributes and convert one into a matrix of receiver attributes
}
