% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_df_function.R
\name{reorder_df}
\alias{reorder_df}
\title{reorder_df}
\usage{
reorder_df(df, col_sort, reorder_data)
}
\arguments{
\item{df}{dataframe to reorder}

\item{col_sort}{column on which the rows will be reordered}

\item{reorder_data}{vector with the new order}
}
\value{
Reordered dataframe
}
\description{
Reorders the rows of one dataframe according to another vector (id vector)
}
\examples{
df <- data.frame(a = letters[1:3],b = LETTERS[4:6],c = 7:9)

reorder_data<-c("c","a","b")

df_new<-reorder_df(df,"a",reorder_data)

df_new

}
