% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_matrix_function.R
\name{reorder_matrix}
\alias{reorder_matrix}
\title{reorder_matrix}
\usage{
reorder_matrix(mat, new_order)
}
\arguments{
\item{mat}{matrix}

\item{new_order}{characters with new order for rows & columns (names)}
}
\value{
reordered matrix
}
\description{
This function reorders matrix rows and columns based on names of another vector
}
\examples{
##Create matrix
M<-replicate(5, rnorm(5))
rownames(M)<-LETTERS[1:5]
colnames(M)<-LETTERS[1:5]

##New order
reorder<-c("E","C","A","B","E")

new_order_matrix<-reorder_matrix(M,reorder)

}
