% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_df_function.R
\name{round_df}
\alias{round_df}
\title{round_df}
\usage{
round_df(x, digits)
}
\arguments{
\item{x}{dataframe}

\item{digits}{digits to round to}
}
\value{
Dataframe with rounded numbers
}
\description{
This function rounds the numeric variables in a dataframe containing numeric and non-numeric data
}
\examples{
##Create dataframe
ID = c("a","b","c","d","e")
Value1 = c(3.445662,6.44566,8.75551,1.114522,1.5551)
Value2 = c(8.2,1.7,6.4,19.45459,10.34524)
df<-data.frame(ID,Value1,Value2)

##Round to 2 digits
rounddf<-round_df(df,2)
}
