% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sender_mat_function.R
\name{sender_mat}
\alias{sender_mat}
\title{sender_mat}
\usage{
sender_mat(DF, attrname)
}
\arguments{
\item{DF}{Dataframe of node attribute}

\item{attrname}{names of the attribute from the dataframe to create the matrix for.}
}
\value{
Sender matrix
}
\description{
This takes a dataframe of node attributes and convert one into a matrix of sender attributes
}
