% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WITSclean_function.R
\name{WITSclean}
\alias{WITSclean}
\title{WITS data clean}
\usage{
WITSclean(CSVfile, YEAR, threshold, cutoff)
}
\arguments{
\item{CSVfile}{WITS csv file}

\item{YEAR}{Year}

\item{threshold}{Apply a threshold - TRUE, Extract the backbone - FALSE}

\item{cutoff}{Threshold - cutoff level, Backbone - significance level}
}
\value{
International Trade Network - igraph object
}
\description{
This function takes trade data downloaded from WITS, cleans it and transforms it into a network.
}
