% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELEnetList_data.R
\docType{data}
\name{ELEnetList}
\alias{ELEnetList}
\title{List of Electical Automtoive Goods Networks (2006-2016)}
\usage{
ELEnetList
}
\description{
List of Electical Automtoive Goods Networks for 2006 - 2016 (list of igraph objects)
}
\keyword{datasets}
