% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ITNdegdist_function.R
\name{ITNdegdist}
\alias{ITNdegdist}
\title{ITN Degree Distribution}
\usage{
ITNdegdist(gs)
}
\arguments{
\item{gs}{International Trade Network - igraph object}
}
\value{
Panel of ITN degree distribution plots
}
\description{
This function plots the ITN (probability) degree distribtuion
}
\examples{
require(igraph)

##Create random International Trade Network (igraph object)
ITN<-erdos.renyi.game(75,0.05,directed = TRUE)

##Plot degree distribution
deg_dist_plot<-ITNdegdist(ITN)

}
