% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ITNplotset_function.R
\name{ITNplotset}
\alias{ITNplotset}
\title{ITN Plots}
\usage{
ITNplotset(gs)
}
\arguments{
\item{gs}{International Trade Network - igraph object}
}
\value{
Panel of ITN plots
}
\description{
This function plots the ITN for a quick inspection
}
\examples{
\donttest{
##Load the network
data(ELEnet16)

##Plot set of network visualisations
ITNplotset(ELEnet16)
}
}
