% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ITNplotset_function.R
\name{ITNplotset}
\alias{ITNplotset}
\title{ITN Plots}
\usage{
ITNplotset(gs)
}
\arguments{
\item{gs}{International Trade Network - igraph object}
}
\value{
Panel of ITN plots
}
\description{
This function creates a panel of four plots of the ITN for a quick inspection.
These include plots: (i) highlighting clusters using the fast greedy algorithm.(ii)node colours for communities detected using the spinglass algorithm.
(iii)nodes coloured by regional partition and (iv)with nodes coloured by regional partition and node size based on outdegree centrality.
}
\examples{
\donttest{
##Load the network
data(ELEnet16)

##Plot set of network visualisations
ITNplotset(ELEnet16)

}
}
