% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ITNproperties_function.R
\name{ITNproperties}
\alias{ITNproperties}
\title{ITN Properties}
\usage{
ITNproperties(gs)
}
\arguments{
\item{gs}{International Trade Network - igraph object}
}
\value{
Table of centrality results (dataframe)
}
\description{
This function calculates network level properties of the International Trade Network (ITN)
}
\examples{
##Load the network
data(ELEnet16)

##Calculate the network properties
ITNPROP<-ITNproperties(ELEnet16)

}
