% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_circle_plot_function.R
\name{region_circle_plot}
\alias{region_circle_plot}
\title{region_circle_plot}
\usage{
region_circle_plot(gs)
}
\arguments{
\item{gs}{igraph ITN object (with attributes added)}
}
\value{
International Trade Network - igraph object
}
\description{
This function creates a chord diagram/circle plot for levels of trade between regional partitions
}
\examples{
##Load graph
data("ELEnet16")

##Create region circle plot
region_circle_plot(ELEnet16)
}
